Refactor project to simplify publishing
This chanage aims at simplify publishing `krun` in crates.io. Since
krun, krun-guest and krun-server are tightly coupled, I don't think it's
worth treating them independently. Refactoring them to be modules within
krun, while specifying them as additional binaries, simplifies
publishing and packaging without significant drawbacks.

We keep krun-sys as an indepenent crate to be published by itself, since
it can potentially be consumed by other programs for linking against
libkrun.
