+++ Cargo.toml
@@ -2,29 +2,3 @@
 resolver = "2"
 members = ["crates/*"]
 
-[workspace.package]
-version = "0.1.0"
-authors = ["Sergio Lopez <slp@redhat.com>", "Teoh Han Hui <teohhanhui@gmail.com>"]
-edition = "2021"
-rust-version = "1.77.0"
-description = "Run programs from your system in a microVM"
-repository = "https://github.com/slp/krun"
-license = "MIT"
-
-[workspace.dependencies]
-anyhow = { version = "1.0.82", default-features = false }
-bindgen = { version = "0.69.4", default-features = false }
-bpaf = { version = "0.9.11", default-features = false }
-byteorder = { version = "1.5.0", default-features = false }
-env_logger = { version = "0.11.3", default-features = false }
-krun-sys = { path = "crates/krun-sys", default-features = false }
-log = { version = "0.4.21", default-features = false }
-nix = { version = "0.28.0", default-features = false }
-rustix = { version = "0.38.34", default-features = false }
-serde = { version = "1.0.203", default-features = false }
-serde_json = { version = "1.0.117", default-features = false }
-tempfile = { version = "3.10.1", default-features = false }
-tokio = { version = "1.38.0", default-features = false }
-tokio-stream = { version = "0.1.15", default-features = false }
-utils = { path = "crates/utils", default-features = false }
-uuid = { version = "1.10.0", default-features = false }
