--- crates/krun-server/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "krun-server"
-version = { workspace = true }
-authors = { workspace = true }
-edition = { workspace = true }
-rust-version = { workspace = true }
-description = { workspace = true }
-repository = { workspace = true }
-license = { workspace = true }
-
-[dependencies]
-anyhow = { workspace = true, features = ["std"] }
-bpaf = { workspace = true, features = [] }
-env_logger = { workspace = true, features = ["auto-color", "humantime", "unstable-kv"] }
-log = { workspace = true, features = ["kv"] }
-serde = { workspace = true, features = [] }
-serde_json = { workspace = true, features = ["std"] }
-tokio = { workspace = true, features = ["io-util", "macros", "net", "process", "rt-multi-thread", "sync"] }
-tokio-stream = { workspace = true, features = ["net", "sync"] }
-utils = { workspace = true, features = [] }
-
-[features]
-default = []
