+++ crates/krun-sys/Cargo.toml
@@ -1,15 +1,14 @@
 [package]
 name = "krun-sys"
 version = "1.9.1"
-edition = { workspace = true }
-rust-version = { workspace = true }
+edition = "2021"
+rust-version = "1.77.0"
 repository = "https://github.com/containers/libkrun"
 license = "Apache-2.0"
 links = "krun"
-publish = false
 
 [build-dependencies]
-bindgen = { workspace = true, features = ["logging"] }
+bindgen = { version = "0.69.4", default-features = false }
 
 [features]
 default = []
