+++ crates/krun/src/lib.rs
@@ -4,3 +4,7 @@ pub mod env;
 pub mod launch;
 pub mod net;
 pub mod types;
+
+pub mod guest;
+pub mod server;
+pub mod utils;
