+++ crates/krun/src/server/server.rs
@@ -13,8 +13,9 @@ use tokio::sync::watch;
 use tokio::task::{JoinError, JoinSet};
 use tokio_stream::wrappers::TcpListenerStream;
 use tokio_stream::StreamExt as _;
-use utils::launch::Launch;
-use utils::stdio::make_stdout_stderr;
+
+use crate::utils::launch::Launch;
+use crate::utils::stdio::make_stdout_stderr;
 
 #[derive(Debug)]
 pub struct Server {
