+++ crates/krun/src/utils/stdio.rs
