--- crates/utils/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "utils"
-authors = { workspace = true }
-edition = { workspace = true }
-rust-version = { workspace = true }
-repository = { workspace = true }
-license = { workspace = true }
-publish = false
-
-[dependencies]
-anyhow = { workspace = true, features = ["std"] }
-serde = { workspace = true, features = ["derive"] }
-uuid = { workspace = true, features = ["std", "v7"] }
-
-[features]
-default = []
